Update Rayleigh Channel.
There is now a vector to store the channel gains called H. It must be empty for awgn. Modulators behavior don't change if H is empty.
It is passed to all the modulators as they need it to demodulate in case of a Rayleigh channel.
Add parameter channl block_fading. If set to NO, channel gain is different for each symbol.
Future implementations will include FRAME to change only on a new frame including interframe and ONETAP excluding interframe

Signed-off-by: Mathieu Léonardon <mathieu.leonardon@u-bordeaux.fr>